+++ include/cry_sum.h
@@ -18,27 +18,40 @@
  */
 
 /**
- * @file    cry.h
- * @brief   CRY Library 
- *
+ * @file    cry_sum.h
+ * @brief   Common checksum algorithms.
  */
 
-#ifndef _CRY_H_
-#define _CRY_H_
+#ifndef _CRY_SUM_H_
+#define _CRY_SUM_H_
 
-#include "cry_version.h"
-#include "cry_base64.h"
-#include "cry_des.h"
-#include "cry_aes.h"
-#include "cry_ciph.h"
-#include "cry_cbc.h"
-#include "cry_gcm.h"
-#include "cry_ctr.h"
-#include "cry_crc.h"
-#include "cry_md5.h"
-#include "cry_sha256.h"
-#include "cry_cmac.h"
-#include "cry_sum.h"
+#include <stddef.h>
 
-#endif /* _CRY_H_ */
+#ifdef __cplusplus
+extern "C"{
+#endif
+
+/**
+ * Trivial checksum.
+ *
+ * @param in    Input buffer.
+ * @param n     Size of input buffer.
+ * @return      Checksum value.
+ */
+unsigned char cry_cs(const unsigned char *in, size_t n);
+
+/**
+ * Longitudinal redundancy check (LRC).
+ *
+ * @param in    Input buffer.
+ * @param n     Size of input buffer.
+ * @return      LRC value.
+ */
+unsigned char cry_lrc(const unsigned char *in, size_t n);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* _CRY_SUM_H_ */